Added regression test for v0 repo without recipes dir

+++ macrotests/repo-regression.sh
@@ -38,6 +38,7 @@ REPO_PATH=$REPO $BOAR ci && { echo "Ci executed ok even though workdir is out of
 REPO_PATH=$REPO $BOAR update || { echo "Couldn't execute update to latest version"; exit 1; }
 REPO_PATH=$REPO $BOAR ci || { echo "Couldn't execute ci"; exit 1; }
 REPO_PATH=$REPO $BOAR verify || { echo "Couldn't verify"; exit 1; }
+test ! -e $REPO/recipes || { echo "recipes dir wasn't deleted"; exit 1; }
 cd $testdir || exit 1
 rm -r regression-boar-daily.11-Jul-2011 || exit 1
 
@@ -60,6 +61,13 @@ tar xzf $TESTHOME/regression-boar-daily.11-Jul-2011.tar.gz || exit 1
 rm $REPO/recovery.txt || exit 1
 REPO_PATH=$REPO $BOAR verify || { echo "Couldn't verify"; exit 1; }
 test -e $REPO/recovery.txt || { echo "recovery.txt wasn't recreated"; exit 1; }
+rm -r regression-boar-daily.11-Jul-2011 || exit 1
+
+echo "--- Test repository without recipes dir"
+# version 0 repos may or may not have a recipes dir
+tar xzf $TESTHOME/regression-boar-daily.11-Jul-2011.tar.gz || exit 1
+rmdir $REPO/recipes || exit 1
+REPO_PATH=$REPO $BOAR verify || { echo "Couldn't verify"; exit 1; }
 rm -r regression-boar-daily.11-Jul-2011 || exit 1
 
 echo "--- Test repo cloning"
